Further fixes.
